+++ src/elements/CardExpirationDateElement.tsx
@@ -25,6 +25,8 @@ interface CardExpirationDateElementProps {
   onBlur?: ElementEventListener<CardExpirationDateElementEvents, 'blur'>;
   onReady?: ElementEventListener<CardExpirationDateElementEvents, 'ready'>;
   onKeyDown?: ElementEventListener<CardExpirationDateElementEvents, 'keydown'>;
+  validateOnChange?: boolean;
+  enableCopy?: boolean;
 }
 
 const CardExpirationDateElementC: FC<
@@ -46,6 +48,8 @@ const CardExpirationDateElementC: FC<
   onBlur,
   onKeyDown,
   elementRef,
+  validateOnChange,
+  enableCopy,
 }) => {
   const wrapperRef = useRef<HTMLDivElement>(null);
   const element = useElement<
@@ -57,12 +61,14 @@ const CardExpirationDateElementC: FC<
     wrapperRef,
     {
       targetId: id,
+      enableCopy,
       style,
       disabled,
       autoComplete,
       'aria-label': ariaLabel,
       placeholder,
       value,
+      validateOnChange,
     },
     bt,
     elementRef