+++ src/elements/CardVerificationCodeElement.tsx
@@ -29,6 +29,8 @@ interface CardVerificationCodeElementProps {
     CardVerificationCodeElementEvents,
     'keydown'
   >;
+  validateOnChange?: boolean;
+  enableCopy?: boolean;
 }
 
 const CardVerificationCodeElementC: FC<
@@ -51,6 +53,8 @@ const CardVerificationCodeElementC: FC<
   onBlur,
   onKeyDown,
   elementRef,
+  validateOnChange,
+  enableCopy,
 }) => {
   const wrapperRef = useRef<HTMLDivElement>(null);
   const element = useElement<
@@ -69,6 +73,8 @@ const CardVerificationCodeElementC: FC<
       placeholder,
       cardBrand,
       value,
+      validateOnChange,
+      enableCopy,
     },
     bt,
     elementRef