+++ test/elements/CardExpirationDateElement.test.tsx
@@ -25,6 +25,8 @@ describe('CardExpirationDateElement', () => {
   let ariaLabel: string;
   let placeholder: string;
   let value: CardExpirationDateValue<'static'>;
+  let validateOnChange: boolean;
+  let enableCopy: boolean;
   let onReady: jest.Mock;
   let onChange: jest.Mock;
   let onFocus: jest.Mock;
@@ -51,6 +53,8 @@ describe('CardExpirationDateElement', () => {
       }),
       year: new Date().getFullYear() + 1,
     };
+    validateOnChange = chance.bool();
+    enableCopy = chance.bool();
     onReady = jest.fn();
     onChange = jest.fn();
     onFocus = jest.fn();
@@ -70,6 +74,7 @@ describe('CardExpirationDateElement', () => {
         aria-label={ariaLabel}
         autoComplete={autoComplete}
         disabled={disabled}
+        enableCopy={enableCopy}
         id={id}
         onBlur={onBlur}
         onChange={onChange}
@@ -79,6 +84,7 @@ describe('CardExpirationDateElement', () => {
         placeholder={placeholder}
         ref={ref}
         style={style}
+        validateOnChange={validateOnChange}
         value={value}
       />
     );
@@ -92,9 +98,11 @@ describe('CardExpirationDateElement', () => {
         targetId: id,
         style,
         disabled,
+        enableCopy,
         autoComplete,
         'aria-label': ariaLabel,
         placeholder,
+        validateOnChange,
         value,
       },
       undefined,