+++ test/elements/CardVerificationCodeElement.test.tsx
@@ -27,6 +27,8 @@ describe('CardVerificationCodeElement', () => {
   let placeholder: string;
   let cardBrand: Brand | string;
   let value: string;
+  let validateOnChange: boolean;
+  let enableCopy: boolean;
   let onReady: jest.Mock;
   let onChange: jest.Mock;
   let onFocus: jest.Mock;
@@ -47,7 +49,8 @@ describe('CardVerificationCodeElement', () => {
     ariaLabel = chance.string();
     placeholder = chance.string();
     value = chance.string();
-
+    validateOnChange = chance.bool();
+    enableCopy = chance.bool();
     cardBrand = chance.pickone<Brand>([...CARD_BRANDS]);
     onReady = jest.fn();
     onChange = jest.fn();
@@ -69,6 +72,7 @@ describe('CardVerificationCodeElement', () => {
         autoComplete={autoComplete}
         cardBrand={cardBrand}
         disabled={disabled}
+        enableCopy={enableCopy}
         id={id}
         onBlur={onBlur}
         onChange={onChange}
@@ -78,6 +82,7 @@ describe('CardVerificationCodeElement', () => {
         placeholder={placeholder}
         ref={ref}
         style={style}
+        validateOnChange={validateOnChange}
         value={value}
       />
     );
@@ -91,11 +96,13 @@ describe('CardVerificationCodeElement', () => {
         targetId: id,
         style,
         disabled,
+        enableCopy,
         autoComplete,
         'aria-label': ariaLabel,
         placeholder,
         cardBrand,
         value,
+        validateOnChange,
       },
       undefined,
       // eslint-disable-next-line unicorn/no-null